+++ src/mmwave/model/mmwave-flex-tti-maxweight-mac-scheduler.cc
@@ -451,7 +451,7 @@ MmWaveFlexTtiMaxWeightMacScheduler::DoSchedUlCqiInfoReq (const struct MmWaveMacS
 {
   NS_LOG_FUNCTION (this);
 
-  uint16_t frameNum = params.m_sfnSf.m_frameNum;
+  uint32_t frameNum = params.m_sfnSf.m_frameNum;
   uint8_t subframeNum =  params.m_sfnSf.m_sfNum;
   uint8_t slotNum = params.m_sfnSf.m_slotNum;
   uint8_t symNum =  params.m_sfnSf.m_symStart;
@@ -751,7 +751,7 @@ unsigned MmWaveFlexTtiMaxWeightMacScheduler::CalcMinTbSizeNumSym (unsigned mcs, 
 void
 MmWaveFlexTtiMaxWeightMacScheduler::DoSchedTriggerReq (const struct MmWaveMacSchedSapProvider::SchedTriggerReqParameters& params)
 {
-  uint16_t frameNum = params.m_snfSf.m_frameNum;
+  uint32_t frameNum = params.m_snfSf.m_frameNum;
   uint8_t sfNum = params.m_snfSf.m_sfNum;
   uint8_t slotNum = params.m_snfSf.m_slotNum;
