+++ src/mmwave/model/mmwave-ue-mac.cc
@@ -635,7 +635,7 @@ std::map<uint32_t, struct MacPduInfo>::iterator MmWaveUeMac::AddToMacPduMap (Dci
   uint8_t slotNum = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) % m_phyMacConfig->GetSlotsPerSubframe ();
   uint8_t deltaSubframe = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) / m_phyMacConfig->GetSlotsPerSubframe ();
   uint8_t sfNum = (m_sfNum + deltaSubframe) % m_phyMacConfig->GetSubframesPerFrame ();
-  uint16_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
+  uint32_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
 
   NS_ASSERT ((slotNum < m_phyMacConfig->GetSlotsPerSubframe ()) && (sfNum < m_phyMacConfig->GetSubframesPerFrame ())
              && (deltaSubframe >= 0) && (slotNum >= 0) && (sfNum >= 0) && (frameNum >= m_frameNum));
@@ -702,7 +702,7 @@ MmWaveUeMac::DoReceiveControlMessage  (Ptr<MmWaveControlMessage> msg)
                     uint8_t slotNum = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) % m_phyMacConfig->GetSlotsPerSubframe ();
                     uint8_t deltaSubframe = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) / m_phyMacConfig->GetSlotsPerSubframe ();
                     uint8_t sfNum = (m_sfNum + deltaSubframe) % m_phyMacConfig->GetSubframesPerFrame ();
-                    uint16_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
+                    uint32_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
 
                     NS_ASSERT ((slotNum < m_phyMacConfig->GetSlotsPerSubframe ()) && (sfNum < m_phyMacConfig->GetSubframesPerFrame ())
                                 && (deltaSubframe >= 0) && (slotNum >= 0) && (sfNum >= 0) && (frameNum >= m_frameNum));
@@ -894,7 +894,7 @@ MmWaveUeMac::DoReceiveControlMessage  (Ptr<MmWaveControlMessage> msg)
                     uint8_t slotNum = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) % m_phyMacConfig->GetSlotsPerSubframe ();
                     uint8_t deltaSubframe = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) / m_phyMacConfig->GetSlotsPerSubframe ();
                     uint8_t sfNum = (m_sfNum + deltaSubframe) % m_phyMacConfig->GetSubframesPerFrame ();
-                    uint16_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
+                    uint32_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
 
                     NS_ASSERT ((slotNum < m_phyMacConfig->GetSlotsPerSubframe ()) && (sfNum < m_phyMacConfig->GetSubframesPerFrame ())
                                 && (deltaSubframe >= 0) && (slotNum >= 0) && (sfNum >= 0) && (frameNum >= m_frameNum));
